refactor(types): 合并重复的 env.d.ts，共用根目录一份全局声明
