+++ apps/system/tsconfig.json
@@ -15,6 +15,7 @@
     "../useMicroApp.ts",
     "../../components/**/*",
     "../../components/**/*.vue",
-    "../../assets/**/*"
+    "../../assets/**/*",
+    "../../env.d.ts"
   ]
 }
